+++ src/core/component/decorators/interface/accessor.ts
@@ -11,7 +11,7 @@ import type { DecoratorFunctionalOptions } from 'core/component/decorators/inter
 
 export interface DecoratorComponentAccessor extends DecoratorFunctionalOptions {
 	/**
-	 * If true, the accessor value will be cached after the first touch.
+	 * If set to true, the accessor value will be cached after the first touch.
 	 * The option is set to true by default if it also provided `dependencies` or the bound accessor matches
 	 * by the name with another prop or field.
 	 * If the option value is passed as `auto` caching will be delegated to the used component library.
@@ -45,7 +45,7 @@ export interface DecoratorComponentAccessor extends DecoratorFunctionalOptions {
 	cache?: ComponentAccessorCacheType;
 
 	/**
-	 * If true, the accessor returns a link to another watchable object.
+	 * If set to true, the accessor returns a link to another watchable object.
 	 * This option allows you to mount external watchable objects to the component.
 	 *
 	 * @example